replaced global settings objects with static objects accessible via getDefaults

+++ lib/core/Globals.h
@@ -9,7 +9,7 @@
 NAMESPACE_SPH_BEGIN
 
 /// Precision used withing the code. Use Float instead of float or double where precision is important.
-using Float = double;
+using Float = float;
 
 /// Integral type used to index arrays (by default).
 using Size = uint32_t;
